Fixed slaves flood if VM start failed
In initial implementation we tried to start VM and created slave node at
the same time. If VM start had failed node was left offline. On the next
provisioner request we tried to start the same VM once again -- and one
more offline node was created. And so on. This commit adds obvious check
for VM was being tried to start.