Fixed VM slaves execution through JNLP
It was impossible to run Jenkins slaves via JNLP because for each new
job VM slave started with the new name. Thus user had to go to VM and
rerun JNLP slave manually. This patch makes VM slave name permanent: it
is formed as "CLOUD_NAME: VM_ID". So user has to run JNLP slave only the
first time and just leave it running. Next time when VM is resumed for
the next job this slave will successfully connect to master
automatically.